add routes and oauth

+++ src/routes/memories.ts
@@ -4,11 +4,21 @@ import { z } from 'zod'
 
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
+
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
-      }
+      },
     })
     return memories.map(memory => {
       return {
@@ -19,7 +29,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramSchema = z.object({
       id: z.string().uuid(),
     })
@@ -32,10 +42,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
 
+    if (!memory.isPublic && memory.userId != request.user.sub) {
+      return reply.status(401).send()
+    }
+
     return memory
   })
 
-  app.post('/memories', async () => {
+  app.post('/memories', async (request) => {
 
     const bodySchema = z.object({
       content: z.string(),
@@ -50,7 +64,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: 'c47d0f84-6e63-454a-9f39-8439a3b038b4'
+        userId: request.user.sub,
       },
 
     })
@@ -59,7 +73,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
 
     const paramSchema = z.object({
       id: z.string().uuid(),
@@ -75,7 +89,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if (memory.userId != request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -92,18 +116,23 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
   })
 
-
-
-
-
-
-  app.delete('/memories/:id', async () => {
+  app.delete('/memories/:id', async (request, reply) => {
 
     const paramSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if (memory.userId != request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
